+++ tests/unit/auths/__init__.py
@@ -1,0 +1,1 @@
+# ~/clientfactory/tests/unit/auths/__init__.py